+++ CHANGELOG.md
@@ -1,6 +1,11 @@
 # Changelog
 
-## Version 0.1 (development)
+
+## Version 0.2
+- refactor DataFrame as BiocFrame
+- implementing slicing methods, tests
+
+## Version 0.1
 
 - Initial creation of DataFrame class
 - Tests